pdf_parser/rootfs/requirements.txt
pdfplumber
tabulate
gspread
gunicorn
pandas
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
